Change format for stashing CellComplex in IFC
Previously the CellComplex was represented as an aggregate of lots of
IfcVirtualElement faces and points. An IfcVirtualElement is not allowed
to have representations, and they were annoyingly visible in the loaded
model.

The CellComplex is now stashed as Tessellated representations of the
building itself, but not visible in BlenderBIM since a footprint
representation gets priority. Cell usages are stashed as annotations,
also attached to the building.